feat: customizable left panel
