+++ packages/jdm-editor/src/components/decision-graph/dg.scss
@@ -128,6 +128,7 @@
         justify-content: flex-start;
         align-items: center;
         height: 36px;
+        user-select: none;
 
         &__actions {
           display: flex;
