+++ packages/jdm-editor/src/components/decision-graph/simulator/simulation.types.ts
